requirements.txt
pymongo>=4.6.0
pytrends>=4.9.2
httpx>=0.25.0
feedparser>=6.0.10
beautifulsoup4>=4.12.2
selectolax>=0.3.17
python-dotenv>=1.0.0
pydantic>=2.5.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
click>=8.1.7

